+++ .gitignore
@@ -1,3 +1,3 @@
-.env
+*.env
 .vscode
 vendor
