+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM denoland/deno:1.20.3
+FROM denoland/deno:1.28.1
 
 EXPOSE 80
 
@@ -7,9 +7,9 @@ WORKDIR /hangman
 USER deno
 
 COPY deps.ts .
-RUN deno cache deps.ts
+RUN deno cache --no-lock deps.ts
 
 COPY . .
-RUN deno cache start.ts
+RUN deno cache --no-lock start.ts
 
-CMD ["run", "--allow-env", "--allow-net", "--cached-only", "--no-check", "start.ts"]
+CMD ["run", "--allow-env", "--allow-net", "--cached-only", "--no-lock", "start.ts"]
