+++ docker-compose.yaml
@@ -1,30 +1,30 @@
-version: '3'
+version: "3"
 services:
   bot:
     build: .
     image: hangman
     container_name: hangman-bot
-    command: deno run --allow-env --allow-net --no-check=remote start.ts
+    command: deno run --allow-env --allow-net --no-lock --check --watch start.ts
     ports:
-      - '80:80'
+      - 80:80
     depends_on:
       - redis
       - mongo
     volumes:
       - .:/hangman
-    env_file: '.env'
+    env_file: dev.env
   mongo:
     image: mongo:5.0.6
     container_name: hangman-mongo
     ports:
-      - '27017:27017'
-    env_file: '.env'
+      - 27017:27017
+    env_file: dev.env
   redis:
     image: redis:6.2.6
     container_name: hangman-redis
     command: redis-server /usr/local/etc/redis/redis.conf
     ports:
-      - '6379:6379'
+      - 6379:6379
     depends_on:
       - mongo
     volumes:
